+++ plugins/catalog-backend-module-bitbucket-server/config.d.ts
@@ -37,24 +37,23 @@ export interface Config {
              */
             filters?: {
               /**
-               * (Optional) Filter for the repository slug.
+               * (Optional) Regular expression filter for the repository slug.
                * @visibility frontend
                */
-              repoSlug?: RegExp;
+              repoSlug?: string;
               /**
-               * (Optional) Filter for the project key.
+               * (Optional) Regular expression filter for the project key.
                * @visibility frontend
                */
-              projectKey?: RegExp;
+              projectKey?: string;
             };
             /**
              * (Optional) TaskScheduleDefinition for the refresh.
              */
             schedule?: TaskScheduleDefinitionConfig;
           }
-        | Record<
-            string,
-            {
+        | {
+            [name: string]: {
               /**
                * (Optional) Path to the catalog file. Default to "/catalog-info.yaml".
                * @visibility frontend
@@ -66,22 +65,22 @@ export interface Config {
                */
               filters?: {
                 /**
-                 * (Optional) Filter for the repository slug.
+                 * (Optional) Regular expression filter for the repository slug.
                  * @visibility frontend
                  */
-                repoSlug?: RegExp;
+                repoSlug?: string;
                 /**
-                 * (Optional) Filter for the project key.
+                 * (Optional) Regular expression filter for the project key.
                  * @visibility frontend
                  */
-                projectKey?: RegExp;
+                projectKey?: string;
               };
               /**
                * (Optional) TaskScheduleDefinition for the refresh.
                */
               schedule?: TaskScheduleDefinitionConfig;
-            }
-          >;
+            };
+          };
     };
   };
 }