+++ plugins/catalog-backend-module-github/src/module/catalogModuleGithubEntityProvider.ts
@@ -34,7 +34,7 @@ export const catalogModuleGithubEntityProvider = createBackendModule({
     env.registerInit({
       deps: {
         catalog: catalogProcessingExtensionPoint,
-        config: coreServices.config,
+        config: coreServices.rootConfig,
         logger: coreServices.logger,
         scheduler: coreServices.scheduler,
       },