+++ packages/techdocs-cli-embedded-app/CHANGELOG.md
@@ -1,5 +1,34 @@
 # techdocs-cli-embedded-app
 
+## 0.2.85-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.12.1-next.2
+  - @backstage/plugin-techdocs@1.6.6-next.2
+  - @backstage/cli@0.22.10-next.1
+  - @backstage/integration-react@1.1.16-next.1
+
+## 0.2.85-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/integration-react@1.1.16-next.1
+  - @backstage/cli@0.22.10-next.1
+  - @backstage/plugin-catalog@1.12.1-next.1
+  - @backstage/plugin-techdocs@1.6.6-next.1
+  - @backstage/app-defaults@1.4.2-next.0
+  - @backstage/catalog-model@1.4.1
+  - @backstage/config@1.0.8
+  - @backstage/core-app-api@1.9.1-next.0
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/test-utils@1.4.2-next.0
+  - @backstage/theme@0.4.1
+  - @backstage/plugin-techdocs-react@1.1.9-next.0
+
 ## 0.2.85-next.0
 
 ### Patch Changes