+++ plugins/azure-devops/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-azure-devops
 
+## 0.3.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.8.1-next.1
+
 ## 0.3.3-next.0
 
 ### Patch Changes