+++ plugins/bitbucket-cloud-common/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-bitbucket-cloud-common
 
+## 0.2.9-next.0
+
+### Patch Changes
+
+- 12a8c94eda8d: Add package repository and homepage metadata
+- Updated dependencies
+  - @backstage/integration@1.5.1
+
 ## 0.2.8
 
 ### Patch Changes