+++ plugins/graphql-voyager/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-graphql-voyager
 
+## 0.1.6-next.2
+
+### Patch Changes
+
+- bb1e1c2b26cc: Fix typo in install instructions.
+
+## 0.1.6-next.1
+
+### Patch Changes
+
+- 12a8c94eda8d: Add package repository and homepage metadata
+- Updated dependencies
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/theme@0.4.1
+
 ## 0.1.6-next.0
 
 ### Patch Changes