+++ plugins/linguist/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @backstage/plugin-linguist
 
+## 0.1.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-linguist-common@0.1.1-next.1
+  - @backstage/plugin-catalog-react@1.8.1-next.1
+
+## 0.1.6-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-linguist-common@0.1.1-next.0
+  - @backstage/catalog-model@1.4.1
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/errors@1.2.1
+  - @backstage/theme@0.4.1
+  - @backstage/plugin-catalog-react@1.8.1-next.0
+
 ## 0.1.6-next.0
 
 ### Patch Changes