+++ plugins/newrelic-dashboard/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @backstage/plugin-newrelic-dashboard
 
+## 0.2.14-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.8.1-next.1
+
+## 0.2.14-next.1
+
+### Patch Changes
+
+- 12a8c94eda8d: Add package repository and homepage metadata
+- Updated dependencies
+  - @backstage/catalog-model@1.4.1
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/errors@1.2.1
+  - @backstage/plugin-catalog-react@1.8.1-next.0
+
 ## 0.2.14-next.0
 
 ### Patch Changes