+++ plugins/scaffolder/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @backstage/plugin-scaffolder
 
+## 1.14.2-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.8.1-next.1
+  - @backstage/plugin-scaffolder-react@1.5.2-next.1
+  - @backstage/integration-react@1.1.16-next.1
+
+## 1.14.2-next.1
+
+### Patch Changes
+
+- 8a0490fb669e: Fix the get entities query in the `MyGroupsPicker` to query the `kind=Group` entities.
+- Updated dependencies
+  - @backstage/integration-react@1.1.16-next.1
+  - @backstage/integration@1.5.1
+  - @backstage/catalog-client@1.4.3
+  - @backstage/catalog-model@1.4.1
+  - @backstage/config@1.0.8
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/errors@1.2.1
+  - @backstage/theme@0.4.1
+  - @backstage/types@1.1.0
+  - @backstage/plugin-catalog-common@1.0.15
+  - @backstage/plugin-catalog-react@1.8.1-next.0
+  - @backstage/plugin-permission-react@0.4.14
+  - @backstage/plugin-scaffolder-common@1.3.2
+  - @backstage/plugin-scaffolder-react@1.5.2-next.0
+
 ## 1.14.2-next.0
 
 ### Patch Changes